Refactor: Extract choosing and click season from watch_media
