+++ src/player/watch_media.rs
@@ -28,33 +28,7 @@ pub async fn watch_media(media: Media, img_mode: bool) -> WebDriverResult<()> {
     driver.goto(url).await?;
 
     if media.url.contains("serie/") {
-        let season_items = driver.find_all(By::Css("div[data-season-id]")).await?;
-
-        let mut season_opts: Vec<String> = Vec::new();
-
-        for season in season_items {
-            season_opts.push(season.inner_html().await?);
-        }
-
-        let season_opt = if season_opts.len() > 1 {
-            choose_season(season_opts).unwrap()
-        } else {
-            season_opts[0].to_string()
-        };
-
-        let season_btn_xpath = format!("//div[text()='{}']", season_opt);
-        let season_element = driver.query(By::XPath(&season_btn_xpath)).first().await?;
-
-        // we execute a js script to not be redirect to other page by the pop up
-        driver
-            .execute(
-                r#"
-            arguments[0].click();
-            "#,
-                vec![season_element.to_json()?],
-            )
-            .await
-            .expect(language.click_season_err);
+        choose_season(&driver).await?;
 
         println!("{}", language.getting_episodes_misc_text);
